use powf

+++ src/threepp/controls/OrbitControls.cpp
@@ -142,7 +142,7 @@ float OrbitControls::getAutoRotationAngle() const {
 
 float OrbitControls::getZoomScale() const {
 
-    return std::pow(0.95f, this->zoomSpeed);
+    return std::powf(0.95f, this->zoomSpeed);
 }
 
 void OrbitControls::rotateLeft(float angle) {
